added module for the river
